Add new message for ffmpeg.js web worker to know the reason for aborting
If an FFmpeg worker currently crashes with an abort, you can find the reason in the console, but this is not programmatically readable. This change adds another message type (`abort') which will send directly before the `exit' message to know the internal reason for the failure. With this information, the calling program code can now make adjustments to avoid causing the same error again.